backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# AI/ML
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
openai==1.6.1
anthropic==0.7.7

# Teams Bot
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
aiohttp==3.9.1

# Diagram Generation
plantuml==0.3.0
Pillow==10.1.0
cairosvg==2.7.1

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
